+++ packages/core/src/odata/common/request-builder/batch/batch-request-serializer.ts
@@ -1,18 +1,24 @@
 /* Copyright (c) 2020 SAP SE or an SAP affiliate company. All rights reserved. */
 
 import voca from 'voca';
+import { ODataRequest } from '../../request/odata-request';
+import { ODataRequestConfig } from '../../request/odata-request-config';
 import { MethodRequestBuilderBase } from '../request-builder-base';
 import { BatchChangeSet } from './batch-change-set';
 import { BatchRequestBuilder } from './batch-request-builder';
-
+import {
+  BatchRequestSerializationOptions,
+  BatchSubRequestPathType
+} from './batch-request-options';
 /**
  * Serialize change set to string.
  * @param changeSet - Change set containing a collection of write operations.
- * @param changeSetBoundary Boundary separating parts in the serialized request.
+ * @param options Request serialization options.
  * @returns The serialized string representation of a change set.
  */
 export function serializeChangeSet(
-  changeSet: BatchChangeSet
+  changeSet: BatchChangeSet,
+  options: BatchRequestSerializationOptions = {}
 ): string | undefined {
   if (changeSet.requests.length) {
     return [
@@ -20,7 +26,7 @@ export function serializeChangeSet(
       '',
       `--${changeSet.boundary}`,
       changeSet.requests
-        .map(request => serializeRequest(request))
+        .map(request => serializeRequest(request, options))
         .join(`\n--${changeSet.boundary}\n`),
       `--${changeSet.boundary}--`
     ].join('\n');
@@ -30,10 +36,17 @@ export function serializeChangeSet(
 /**
  * Serialize a multipart request to string.
  * @param request One of [[GetAllRequestBuilder | getAll]], [[GetByKeyRequestBuilder | getByKey]], [[CreateRequestBuilder | create]], [[UpdateRequestBuilder | update]] or [[DeleteRequestBuilder | delete]] request builder.
+ * @param options Request serialization options.
  * @returns The serialized string representation of a multipart request, including the multipart headers.
  */
-export function serializeRequest(request: MethodRequestBuilderBase) {
-  const odataRequest = request.build();
+export function serializeRequest(
+  request: MethodRequestBuilderBase,
+  options: BatchRequestSerializationOptions = {}
+): string {
+  const odataRequest = new ODataRequest(
+    request.requestConfig,
+    options.destination
+  );
   const headers = {
     ...odataRequest.defaultHeaders(),
     ...odataRequest.eTagHeaders(),
@@ -42,16 +55,33 @@ export function serializeRequest(request: MethodRequestBuilderBase) {
   const requestHeaders = Object.entries(headers).map(
     ([key, value]) => `${voca.titleCase(key)}: ${value}`
   );
+
   return [
     'Content-Type: application/http',
     'Content-Transfer-Encoding: binary',
     '',
-    `${request.requestConfig.method.toUpperCase()} /${request.relativeUrl()} HTTP/1.1`,
+    `${request.requestConfig.method.toUpperCase()} ${encodeURI(
+      getUrl(odataRequest, options.subRequestPathType)
+    )} HTTP/1.1`,
     ...(requestHeaders.length ? requestHeaders : ['']),
     '',
     ...getPayload(request),
     ''
   ].join('\n');
+}
+
+function getUrl<ConfigT extends ODataRequestConfig>(
+  request: ODataRequest<ConfigT>,
+  subRequestPathType?: BatchSubRequestPathType
+): string {
+  switch (subRequestPathType) {
+    case 'absolute':
+      return request.url();
+    case 'relativeToEntity':
+      return `/${request.relativeUrl(false)}`;
+    default:
+      return `/${request.relativeUrl()}`;
+  }
 }
 
 function getPayload(request: MethodRequestBuilderBase): string[] {
@@ -60,17 +90,31 @@ function getPayload(request: MethodRequestBuilderBase): string[] {
     : [];
 }
 
+function validateOptions(options: BatchRequestSerializationOptions): void {
+  // This should never happen. Can only occur if requestbuilder.build() was called which will be removed.
+  if (options.subRequestPathType === 'absolute' && !options.destination?.url) {
+    throw new Error(
+      "Cannot serialize batch request. Invalid destination provided for sub request path type 'absolute'"
+    );
+  }
+}
+
 /**
  * Serialize a batch request to string. This is used for the batch request payload when executing the request.
  * @param request - Batch request to serialize.
+ * @param options Request serialization options.
  * @returns String representation of the batch request.
  */
-export function serializeBatchRequest(request: BatchRequestBuilder): string {
+export function serializeBatchRequest(
+  request: BatchRequestBuilder,
+  options: BatchRequestSerializationOptions = {}
+): string {
+  validateOptions(options);
   const serializedSubRequests = request.requests
     .map(subRequest =>
       subRequest instanceof MethodRequestBuilderBase
-        ? serializeRequest(subRequest)
-        : serializeChangeSet(subRequest)
+        ? serializeRequest(subRequest, options)
+        : serializeChangeSet(subRequest, options)
     )
     .filter(validRequest => !!validRequest)
     .join(`\n--${request.requestConfig.boundary}\n`);