+++ packages/util/src/equal.ts
@@ -1,0 +1,31 @@
+export function equalObjects(
+  obj1: Record<string, any>,
+  obj2: Record<string, any>
+): boolean {
+  const keys1 = Object.keys(obj1);
+  return (
+    Object.keys(obj1).length === Object.keys(obj2).length &&
+    keys1.every(key => equal(obj1[key], obj2[key]))
+  );
+}
+
+export function equal<T>(item1: T, item2: T): boolean {
+  if (Array.isArray(item1) && Array.isArray(item2)) {
+    return equalArrays(item1, item2);
+  }
+  if (
+    typeof item1 === 'object' &&
+    typeof item2 === 'object' &&
+    !(item1 === null && item2 === null)
+  ) {
+    return equalObjects(item1, item2);
+  }
+  return item1 === item2;
+}
+
+export function equalArrays<T>(arr1: T[], arr2: T[]): boolean {
+  return (
+    arr1.length === arr2.length &&
+    arr1.every((item1, i) => equal(item1, arr2[i]))
+  );
+}